COMM: fix local id size for communicators to be 32 bits

+++ src/MPC_Lowcomm/src/communicator.c
@@ -915,7 +915,7 @@ static inline void __communicator_id_factory_init(void)
 	int process_count = mpc_common_get_process_count();
 
 	/* We have the whole range except two */
-	uint16_t global_dynamic = ( ( (uint16_t)-1) - COMM_ID_TO_SKIP);
+	uint32_t global_dynamic = ( ( (uint32_t)-1) - COMM_ID_TO_SKIP);
 
 	/* Ensure it can be divided evenly */
 	if(global_dynamic % process_count)
